refactor: Refactorización de actividad preguntados. #89
