+++ src/activity/hooks/useCreatePreguntados.ts
@@ -1,11 +1,11 @@
 import { useContext } from "react";
 import { PreguntadosContext } from "../contexts/preguntadosContext/PreguntadosContext";
 
-export const usePreguntados = () => {
+export const useCreatePreguntados = () => {
   const context = useContext(PreguntadosContext);
   if (!context) {
     throw new Error(
-      "usePreguntados must be used within an PreguntadosProvider"
+      "useCreatePreguntados must be used within an CreatePreguntadosProvider"
     );
   }
   return context;
